refactor: update imports to use shared.clients for GeminiInteractionClient and implement polling dispatcher for deep research tasks
