+++ chat_shell/chat_shell/api/v1/deep_research.py
@@ -19,7 +19,7 @@ from fastapi import APIRouter, HTTPException
 from fastapi.responses import StreamingResponse
 from pydantic import BaseModel, Field
 
-from chat_shell.clients.gemini_interaction import (
+from shared.clients.gemini_interaction import (
     GeminiInteractionClient,
     GeminiInteractionError,
 )
