+++ shared/clients/gemini_interaction.py
